fix(pc-port): Resolve the eight game-code link failures
All eight gone, gate unmoved: DOL 306526d9, 7249 / 80.66999%. pclink.py reports
0 under "game code" and 73 unresolved overall.

Six were not missing functions. They were defined with a different type than the
calling unit declared:

  zEntPlayerDyingInGoo   S32 defined / U8 declared
  xSndStreamReady        U32 / U8
  NPCC_LineHitsBound     U32 / S32
  zMenuCardCheckStartup  bool / S32
  LERP (4-arg)           xVec3* / void
  menu_fmv_played        bool / U8

PowerPC CodeWarrior does not encode a return type in a mangled name and MSVC
does, so retail gets one symbol out of each pair and a host build gets two.

The previous commit said "one side is wrong about what retail returns". That was
wrong, and the evidence was already here. Retail had these splits itself and at
least two are load-bearing: zNPCTypeBossPatrick.cpp records that agreeing with
the definition costs a whole function -- 70/71 to 69/71, the 6040-byte
zNPCBPatrick::Process, which tests only the low byte -- and ends "Do not 'fix'
it". zMain.cpp records that retail's zMain.o assigns zMenuCardCheckStartup's
result with a plain `mr`, which only an int-typed declaration produces.

So each declaration is guarded rather than corrected: #ifdef PLATFORM_PC takes
the definition's type, #else keeps retail's. The console's preprocessed output
is unchanged by construction; the gate was run anyway.

The other two were a real portability defect. xBoulderGenerator_Init and
bungee_state::load each declared a parameter size_t in their header and defined
it unsigned long. types.h typedefs size_t to unsigned long for the console so
the two agree there; on 32-bit Windows size_t is unsigned int, and the two are
distinct types for overload resolution at the same width. Both definitions say
size_t now, which is character-for-character the same declaration on the console.

LINKING.md corrected, including the claim it got wrong.

+++ src/SB/Core/x/xEntBoulder.cpp
@@ -1077,7 +1077,13 @@ static void RecurseChild(xBase* child, xEntBoulder** boulList, S32& currBoul)
     }
 }
 
-void xBoulderGenerator_Init(xBase& data, xDynAsset& asset, unsigned long)
+// size_t, not unsigned long, to agree with the declaration in the header.
+// include/types.h typedefs size_t to unsigned long for the console, so this is
+// character-for-character the same type there and the mangled name does not
+// move. On 32-bit Windows size_t is unsigned int, and unsigned int and
+// unsigned long are distinct types for overload resolution even at the same
+// width -- so the header declared one function and this defined another.
+void xBoulderGenerator_Init(xBase& data, xDynAsset& asset, size_t)
 {
     xBoulderGenerator_Init((xBoulderGenerator*)&data, (xBoulderGeneratorAsset*)&asset);
 }
